Refactor and update song logic
Now handles songs from multiple files and sorts on time listened
