+++ src/processData.ts
@@ -9,15 +9,15 @@ function ProcessContents(file: File): Promise<Song[]> {
                const contents = event?.target?.result as string;
                const jsonData = JSON.parse(contents);
                const songs: Song[] = jsonData.map((data: any) => DataToSong(data));
-               const cleanSongs = CleanSongs(songs);
+               //const cleanSongs = CleanSongs(songs);
                
-               for (var i = 0; i < 50; i++) {
+               /*for (var i = 0; i < 50; i++) {
                    console.log(cleanSongs[i].uri);
                    console.log(cleanSongs[i].name);
                    console.log(cleanSongs[i].timeListened);
-               }
+               }*/
                
-               resolve(cleanSongs);
+               resolve(songs);
            } catch (error) {
                reject(error);
            }
@@ -29,6 +29,7 @@ function ProcessContents(file: File): Promise<Song[]> {
    });
 }
 
+
 function DataToSong(data : any) {
    return {
       uri: data.spotify_track_uri,
@@ -37,27 +38,6 @@ function DataToSong(data : any) {
       album: data.master_metadata_album_album_name,
       timeListened: (+data.ms_played)/1000
    }
-}
-
-function CleanSongs(songs : Song[]) : Song[] {
-   const cleanSongs : Song[] = [];
-   for (const song of songs) {
-      if (song.name === null || song.timeListened == null || song.timeListened < 20) {
-         continue;
-      }
-      var songFound = false;
-      for (const cleanedSong of cleanSongs) {
-         if (cleanedSong.name === song.name && cleanedSong.artist === song.artist) {
-            songFound = true;
-            cleanedSong.timeListened = cleanedSong.timeListened +song.timeListened;
-            break;
-         }
-      }
-      if (!songFound) {
-         cleanSongs.push(song);
-      }
-   }
-   return cleanSongs;
 }
 
 export default ProcessContents
